Set default mode for Trip
When performing a track call after a trip has been started the track call retreives the tripOptions from storage. The track call requires the 'mode' to be set despite it being optional from a trip perspective. We need to set the default mode into the storage to ensure /track works.